blocks: Don't write None for media blocks if files are missing
Problem occurs when primary-source media files that
Image/Video/Document blocks point to are missing during migration.
Wagtail chokes when instantiating StreamBlocks if a None
is written to the image|video|document` value in
encyclopedia.blocks.(Image|Video|DocumentBlock).block_from_source
It would rather the key/value pair just be missing.
The fix is to just not write a(n) `image|video|document` value
to the block if the media object is missing.