migration: Add media blocks, set article body, then attach files
For whatever reason, I get a serialization error when I try
to add primary source media files to the Image/Video/DocumentBlocks
before all the StreamBlocks are added to article.body.

So instead the primary-source info is given to one function which
creates the media Blocks, and then, after article.body is populated
the source info is given to the second function which adds the
file objects.